requirements/requirements.txt
openai
langchain-openai
langchain-cli
langchain-core
langchain-community
sentence-transformers
pypdf
pysqlite3-binary>=0.5.1
langchain-chroma
chromadb
tiktoken
python-dotenv
streamlit
bcrypt==4.1.2